attempt to use OOTB service connection for S3
