+++ aws-s3-project/src/test/java/com/learning/awspring/TestS3Application.java
@@ -1,11 +1,14 @@
 package com.learning.awspring;
 
-import com.learning.awspring.common.ContainerConfig;
+import com.learning.awspring.common.LocalStackContainerConfig;
+import com.learning.awspring.common.SQLContainerConfig;
 import org.springframework.boot.SpringApplication;
 
 public class TestS3Application {
 
     public static void main(String[] args) {
-        SpringApplication.from(S3Application::main).with(ContainerConfig.class).run(args);
+        SpringApplication.from(S3Application::main)
+                .with(SQLContainerConfig.class, LocalStackContainerConfig.class)
+                .run(args);
     }
 }
