+++ aws-s3-project/src/test/java/com/learning/awspring/common/SQLContainerConfig.java
@@ -1,0 +1,34 @@
+package com.learning.awspring.common;
+
+import org.springframework.boot.devtools.restart.RestartScope;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
+import org.springframework.context.annotation.Bean;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.utility.DockerImageName;
+
+@TestConfiguration(proxyBeanMethods = false)
+public class SQLContainerConfig {
+
+    @Bean
+    @ServiceConnection
+    @RestartScope
+    PostgreSQLContainer<?> postgresContainer() {
+        return new PostgreSQLContainer<>(DockerImageName.parse("postgres:17.2-alpine"))
+                .withReuse(true);
+    }
+
+    // @Bean
+    // DynamicPropertyRegistrar dynamicPropertyRegistrar(LocalStackContainer localStackContainer) {
+    //     return registry -> {
+    //         registry.add(
+    //                 "spring.cloud.aws.credentials.access-key",
+    // localStackContainer::getAccessKey);
+    //         registry.add(
+    //                 "spring.cloud.aws.credentials.secret-key",
+    // localStackContainer::getSecretKey);
+    //         registry.add("spring.cloud.aws.region.static", localStackContainer::getRegion);
+    //         registry.add("spring.cloud.aws.endpoint", localStackContainer::getEndpoint);
+    //     };
+    // }
+}
