+++ aws-s3-project/src/test/java/com/learning/awspring/repository/SchemaValidationTest.java
@@ -2,18 +2,18 @@ package com.learning.awspring.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import com.learning.awspring.common.SQLContainerConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import javax.sql.DataSource;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.context.annotation.Import;
 
-@DataJpaTest(
-        properties = {
-            "spring.jpa.hibernate.ddl-auto=validate",
-            "spring.test.database.replace=none",
-            "spring.datasource.url=jdbc:tc:postgresql:16.2-alpine:///db"
-        })
+@DataJpaTest(properties = {"spring.jpa.hibernate.ddl-auto=validate"})
+@AutoConfigureTestDatabase
+@Import(SQLContainerConfig.class)
 class SchemaValidationTest {
 
     @Autowired private DataSource dataSource;
